+++ lib/src/tooltip_widget.dart
@@ -21,19 +21,22 @@
  */
 
 import 'dart:math';
+import 'dart:ui';
 
 import 'package:flutter/material.dart';
 
 import 'enum.dart';
+import 'extension.dart';
 import 'get_position.dart';
 import 'measure_size.dart';
+import 'widget/tooltip_slide_transition.dart';
 
 const _kDefaultPaddingFromParent = 14.0;
 
 class ToolTipWidget extends StatefulWidget {
   final GetPosition? position;
   final Offset? offset;
-  final Size? screenSize;
+  final Size screenSize;
   final String? title;
   final TextAlign? titleAlignment;
   final String? description;
@@ -59,9 +62,13 @@ class ToolTipWidget extends StatefulWidget {
   final TooltipPosition? tooltipPosition;
   final EdgeInsets? titlePadding;
   final EdgeInsets? descriptionPadding;
+  final TextDirection? titleTextDirection;
+  final TextDirection? descriptionTextDirection;
+  final double toolTipSlideEndDistance;
+  final Offset? tooltipOffset;
 
   const ToolTipWidget({
-    Key? key,
+    super.key,
     required this.position,
     required this.offset,
     required this.screenSize,
@@ -90,7 +97,11 @@ class ToolTipWidget extends StatefulWidget {
     this.tooltipPosition,
     this.titlePadding,
     this.descriptionPadding,
-  }) : super(key: key);
+    this.titleTextDirection,
+    this.descriptionTextDirection,
+    this.toolTipSlideEndDistance = 7,
+    this.tooltipOffset,
+  });
 
   @override
   State<ToolTipWidget> createState() => _ToolTipWidgetState();
@@ -118,10 +129,20 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
         position.dy + ((widget.position?.getHeight() ?? 0) / 2);
     final topPosition = position.dy - ((widget.position?.getHeight() ?? 0) / 2);
     final hasSpaceInTop = topPosition >= height;
+    // TODO: need to update for flutter version > 3.8.X
+    // ignore: deprecated_member_use
+    final EdgeInsets viewInsets = EdgeInsets.fromWindowPadding(
+      // ignore: deprecated_member_use
+      ambiguate(WidgetsBinding.instance)?.window.viewInsets ??
+          // ignore: deprecated_member_use
+          WindowPadding.zero,
+      // ignore: deprecated_member_use
+      ambiguate(WidgetsBinding.instance)?.window.devicePixelRatio ?? 1,
+    );
+    final double actualVisibleScreenHeight =
+        widget.screenSize.height - viewInsets.bottom;
     final hasSpaceInBottom =
-        ((widget.screenSize?.height ?? MediaQuery.of(context).size.height) -
-                bottomPosition) >=
-            height;
+        (actualVisibleScreenHeight - bottomPosition) >= height;
     return widget.tooltipPosition ??
         (hasSpaceInTop && !hasSpaceInBottom
             ? TooltipPosition.top
@@ -132,12 +153,16 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
     final titleStyle = widget.titleTextStyle ??
         Theme.of(context)
             .textTheme
-            .headline6!
+            // TODO: replace once support for 3.1.0 and above is provided.
+            // ignore: deprecated_member_use
+            .titleLarge!
             .merge(TextStyle(color: widget.textColor));
     final descriptionStyle = widget.descTextStyle ??
         Theme.of(context)
             .textTheme
-            .subtitle2!
+            // TODO: replace once support for 3.1.0 and above is provided.
+            // ignore: deprecated_member_use
+            .titleSmall!
             .merge(TextStyle(color: widget.textColor));
     final titleLength = widget.title == null
         ? 0
@@ -154,8 +179,8 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
             (widget.descriptionPadding?.right ?? 0) +
             (widget.descriptionPadding?.left ?? 0));
     var maxTextWidth = max(titleLength, descriptionLength);
-    if (maxTextWidth > widget.screenSize!.width - tooltipScreenEdgePadding) {
-      tooltipWidth = widget.screenSize!.width - tooltipScreenEdgePadding;
+    if (maxTextWidth > widget.screenSize.width - tooltipScreenEdgePadding) {
+      tooltipWidth = widget.screenSize.width - tooltipScreenEdgePadding;
     } else {
       tooltipWidth = maxTextWidth + tooltipTextPadding;
     }
@@ -166,12 +191,12 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
       final width =
           widget.container != null ? _customContainerWidth.value : tooltipWidth;
       double leftPositionValue = widget.position!.getCenter() - (width * 0.5);
-      if ((leftPositionValue + width) > MediaQuery.of(context).size.width) {
+      if ((leftPositionValue + width) > widget.screenSize.width) {
         return null;
       } else if ((leftPositionValue) < _kDefaultPaddingFromParent) {
         return _kDefaultPaddingFromParent;
       } else {
-        return leftPositionValue;
+        return leftPositionValue + (widget.tooltipOffset?.dx ?? 0);
       }
     }
     return null;
@@ -183,10 +208,12 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
           widget.container != null ? _customContainerWidth.value : tooltipWidth;
 
       final left = _getLeft();
-      if (left == null || (left + width) > MediaQuery.of(context).size.width) {
-        final rightPosition = widget.position!.getCenter() + (width * 0.5);
+      if (left == null || (left + width) > widget.screenSize.width) {
+        final rightPosition = widget.position!.getCenter() +
+            (width * 0.5) +
+            (widget.tooltipOffset?.dx ?? 0);
 
-        return (rightPosition + width) > MediaQuery.of(context).size.width
+        return (rightPosition + width) > widget.screenSize.width
             ? _kDefaultPaddingFromParent
             : null;
       } else {
@@ -198,8 +225,8 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
 
   double _getSpace() {
     var space = widget.position!.getCenter() - (widget.contentWidth! / 2);
-    if (space + widget.contentWidth! > widget.screenSize!.width) {
-      space = widget.screenSize!.width - widget.contentWidth! - 8;
+    if (space + widget.contentWidth! > widget.screenSize.width) {
+      space = widget.screenSize.width - widget.contentWidth! - 8;
     } else if (space < (widget.contentWidth! / 2)) {
       space = 16;
     }
@@ -212,7 +239,7 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
         ? 0
         : (widget.position!.getCenter() - calculatedLeft);
     var right = _getLeft() == null
-        ? (MediaQuery.of(context).size.width - widget.position!.getCenter()) -
+        ? (widget.screenSize.width - widget.position!.getCenter()) -
             (_getRight() ?? 0)
         : 0;
     final containerWidth =
@@ -240,7 +267,7 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance.addPostFrameCallback((_) {
+    ambiguate(WidgetsBinding.instance)?.addPostFrameCallback((_) {
       if (widget.container != null &&
           _customContainerKey.currentContext != null &&
           _customContainerKey.currentContext?.size != null) {
@@ -300,8 +327,14 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
 
   @override
   void didChangeDependencies() {
-    _getTooltipWidth();
     super.didChangeDependencies();
+    _getTooltipWidth();
+  }
+
+  @override
+  void didUpdateWidget(covariant ToolTipWidget oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    _getTooltipWidth();
   }
 
   @override
@@ -329,7 +362,7 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
         contentOffsetMultiplier.clamp(-1.0, 0.0);
 
     var paddingTop = isArrowUp ? 22.0 : 0.0;
-    var paddingBottom = isArrowUp ? 0.0 : 27.0;
+    var paddingBottom = isArrowUp ? 0.0 : 22.0;
 
     if (!widget.showArrow) {
       paddingTop = 10;
@@ -357,10 +390,13 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
               ),
           child: FractionalTranslation(
             translation: Offset(0.0, contentFractionalOffset as double),
-            child: SlideTransition(
+            child: ToolTipSlideTransition(
               position: Tween<Offset>(
-                begin: Offset(0.0, contentFractionalOffset / 10),
-                end: const Offset(0.0, 0.100),
+                begin: Offset.zero,
+                end: Offset(
+                  0,
+                  widget.toolTipSlideEndDistance * contentOffsetMultiplier,
+                ),
               ).animate(_movingAnimation),
               child: Material(
                 type: MaterialType.transparency,
@@ -421,10 +457,14 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
                                       child: Text(
                                         widget.title!,
                                         textAlign: widget.titleAlignment,
+                                        textDirection:
+                                            widget.titleTextDirection,
                                         style: widget.titleTextStyle ??
                                             Theme.of(context)
                                                 .textTheme
-                                                .headline6!
+                                                // TODO: replace once support for 3.1.0 and above is provided.
+                                                // ignore: deprecated_member_use
+                                                .titleLarge!
                                                 .merge(
                                                   TextStyle(
                                                     color: widget.textColor,
@@ -438,10 +478,14 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
                                     child: Text(
                                       widget.description!,
                                       textAlign: widget.descriptionAlignment,
+                                      textDirection:
+                                          widget.descriptionTextDirection,
                                       style: widget.descTextStyle ??
                                           Theme.of(context)
                                               .textTheme
-                                              .subtitle2!
+                                              // TODO: replace once support for 3.1.0 and above is provided.
+                                              // ignore: deprecated_member_use
+                                              .titleSmall!
                                               .merge(
                                                 TextStyle(
                                                   color: widget.textColor,
@@ -468,15 +512,16 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
       children: <Widget>[
         Positioned(
           left: _getSpace(),
-          top: contentY - 10,
+          top: contentY - (10 * contentOffsetMultiplier),
           child: FractionalTranslation(
             translation: Offset(0.0, contentFractionalOffset as double),
-            child: SlideTransition(
+            child: ToolTipSlideTransition(
               position: Tween<Offset>(
-                begin: Offset(0.0, contentFractionalOffset / 10),
-                end: !widget.showArrow && !isArrowUp
-                    ? const Offset(0.0, 0.0)
-                    : const Offset(0.0, 0.100),
+                begin: Offset.zero,
+                end: Offset(
+                  0,
+                  widget.toolTipSlideEndDistance * contentOffsetMultiplier,
+                ),
               ).animate(_movingAnimation),
               child: Material(
                 color: Colors.transparent,
@@ -485,6 +530,7 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
                   child: Container(
                     padding: EdgeInsets.only(
                       top: paddingTop,
+                      bottom: paddingBottom,
                     ),
                     color: Colors.transparent,
                     child: Center(
@@ -513,6 +559,8 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
     final textPainter = TextPainter(
       text: TextSpan(text: text, style: style),
       maxLines: 1,
+      // TODO: replace this once we support sdk v3.12.
+      // ignore: deprecated_member_use
       textScaleFactor: MediaQuery.of(context).textScaleFactor,
       textDirection: TextDirection.ltr,
     )..layout();
@@ -527,7 +575,7 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
 
   double? _getArrowRight(double arrowWidth) {
     if (_getLeft() != null) return null;
-    return (MediaQuery.of(context).size.width - widget.position!.getCenter()) -
+    return (widget.screenSize.width - widget.position!.getCenter()) -
         (_getRight() ?? 0) -
         (arrowWidth / 2);
   }